tests/wpt/tests/css/css-overflow/parsing/scroll-axis-lock-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow: parsing scroll-axis-lock property invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#scroll-axis-lock">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<div id="target"></div>
<script>
  test_invalid_value('scroll-axis-lock', '10');
  test_invalid_value('scroll-axis-lock', 'true');
  test_invalid_value('scroll-axis-lock', 'default');
  test_invalid_value('scroll-axis-lock', 'all');
  test_invalid_value('scroll-axis-lock', 'auto, none');
  test_invalid_value('scroll-axis-lock', 'always');
  test_invalid_value('scroll-axis-lock', 'both');
</script>
